Port to 1.21.11: compile fixes wave 1-5 (records, Matrix3x2fStack, GlStateManager, input/packet APIs, GUI states, mixins)

+++ src/main/java/thunder/hack/gui/font/FontRenderer.java
@@ -184,7 +184,7 @@ public class FontRenderer implements Closeable {
         Matrix3x2fStack stack = context.getMatrices();
         stack.pushMatrix();
         y -= 3f;
-        stack.translate(roundToDecimal(x, 1), roundToDecimal(y, 1));
+        stack.translate((float) roundToDecimal(x, 1), (float) roundToDecimal(y, 1));
         stack.scale(1f / this.scaleMul, 1f / this.scaleMul);
 
         GuiRenderState state = ((IDrawContextAccessor) context).thunderhack$getRenderState();
